separate openai modules
