+++ src/langrila/openai/llm/chat.py
@@ -2,23 +2,25 @@ import asyncio
 from copy import deepcopy
 from typing import Any, AsyncGenerator, Generator, Optional
 
-import numpy as np
 from openai import AsyncAzureOpenAI, AsyncOpenAI, AzureOpenAI, OpenAI
 from PIL import Image
 
-from ..base import (
+from ...base import (
     BaseChatModule,
     BaseConversationLengthAdjuster,
     BaseConversationMemory,
     BaseFilter,
-    BaseModule,
+    BaseMessage,
 )
+from ...llm_wrapper import ChatWrapperModule
+from ...mixin import ConversationMixin, FilterMixin
+from ...result import CompletionResults
+from ...usage import Usage
+from ...utils import make_batch
 from ..conversation_adjuster.truncate import OldConversationTruncationModule
-from ..message import Message
+from ..message import OpenAIMessage
 from ..model_config import _OLDER_MODEL_CONFIG, _VISION_MODEL, MODEL_CONFIG, MODEL_POINT
-from ..result import CompletionResults
-from ..usage import Usage
-from ..utils import get_async_client, get_client, get_n_tokens, get_token_limit, make_batch
+from ..openai_utils import get_async_client, get_client, get_n_tokens, get_token_limit
 
 
 def completion(
@@ -85,7 +87,7 @@ async def acompletion(
     return await client.chat.completions.create(**params)
 
 
-class ChatCoreModule(BaseChatModule):
+class OpenAIChatCoreModule(BaseChatModule):
     def __init__(
         self,
         api_key_env_name: str,
@@ -342,7 +344,7 @@ class ChatCoreModule(BaseChatModule):
                         )
 
 
-class OpenAIChatModule(BaseChatModule):
+class OpenAIChatModule(ChatWrapperModule):
     def __init__(
         self,
         api_key_env_name: str,
@@ -378,7 +380,7 @@ class OpenAIChatModule(BaseChatModule):
         ), f"max_tokens({max_tokens}) + context_length({context_length}) must be less than or equal to the token limit of the model ({token_lim})."
         assert context_length > 0, "context_length must be positive."
 
-        self.chat_model = ChatCoreModule(
+        chat_model = OpenAIChatCoreModule(
             api_key_env_name=api_key_env_name,
             organization_id_env_name=organization_id_env_name,
             model_name=model_name,
@@ -393,153 +395,21 @@ class OpenAIChatModule(BaseChatModule):
             response_format=response_format,
         )
 
-        self.conversation_length_adjuster = (
+        conversation_length_adjuster = (
             OldConversationTruncationModule(model_name=model_name, context_length=context_length)
             if conversation_length_adjuster is None
             else conversation_length_adjuster
         )
-        self.conversation_memory = conversation_memory
-        self.content_filter = content_filter
 
-    def run(
-        self,
-        prompt: str,
-        images: Image.Image | bytes | list[Image.Image | bytes] | None = None,
-        init_conversation: Optional[list[dict[str, str]]] = None,
-        image_resolution: str = "low",
-    ) -> CompletionResults:
-        messages = self.load_conversation()
-
-        if isinstance(init_conversation, list) and len(messages) == 0:
-            messages.extend(init_conversation)
-
-        messages.append(
-            Message(content=prompt, images=images, image_resolution=image_resolution).as_user
+        super().__init__(
+            chat_model=chat_model,
+            conversation_memory=conversation_memory,
+            content_filter=content_filter,
+            conversation_length_adjuster=conversation_length_adjuster,
         )
 
-        if self.content_filter is not None:
-            messages = self.content_filter.apply(messages)
-
-        messages_adjusted = self.conversation_length_adjuster.run(messages)
-
-        response = self.chat_model.run(messages_adjusted)
-
-        if self.content_filter is not None:
-            response.message = self.restore_content_filter([response.message])[0]
-
-        messages.append(response.message)
-
-        if self.conversation_memory is not None:
-            self.save_conversation(messages)
-
-        return response
-
-    def stream(
-        self,
-        prompt: str,
-        images: Image.Image | bytes | list[Image.Image | bytes] | None = None,
-        init_conversation: Optional[list[dict[str, str]]] = None,
-        image_resolution: str = "low",
-    ) -> Generator[CompletionResults, None, None]:
-        messages = self.load_conversation()
-
-        if isinstance(init_conversation, list) and len(messages) == 0:
-            messages.extend(init_conversation)
-
-        messages.append(
-            Message(content=prompt, images=images, image_resolution=image_resolution).as_user
-        )
-
-        if self.content_filter is not None:
-            messages = self.apply_content_filter(messages)
-
-        messages_adjusted = self.conversation_length_adjuster.run(messages)
-
-        response = self.chat_model.stream(messages_adjusted)
-
-        for chunk in response:
-            if isinstance(chunk, CompletionResults):
-                if self.content_filter is not None:
-                    chunk.message = self.restore_content_filter([chunk.message])[0]
-
-                messages.append(chunk.message)
-
-                if self.conversation_memory is not None:
-                    self.save_conversation(messages)
-
-                yield chunk
-            else:
-                raise AssertionError
-
-    async def astream(
-        self,
-        prompt: str,
-        images: Image.Image | bytes | list[Image.Image | bytes] | None = None,
-        init_conversation: Optional[list[dict[str, str]]] = None,
-        image_resolution: str = "low",
-    ) -> AsyncGenerator[CompletionResults, None]:
-        messages = self.load_conversation()
-
-        if isinstance(init_conversation, list) and len(messages) == 0:
-            messages.extend(init_conversation)
-
-        messages.append(
-            Message(content=prompt, images=images, image_resolution=image_resolution).as_user
-        )
-
-        if self.content_filter is not None:
-            messages = self.apply_content_filter(messages)
-
-        messages_adjusted = self.conversation_length_adjuster.run(messages)
-
-        response = self.chat_model.astream(messages_adjusted)
-
-        async for chunk in response:
-            if isinstance(chunk, CompletionResults):
-                if self.content_filter is not None:
-                    chunk.message = self.restore_content_filter([chunk.message])[0]
-
-                messages.append(chunk.message)
-
-                if self.conversation_memory is not None:
-                    self.save_conversation(messages)
-
-                yield chunk
-            else:
-                raise AssertionError
-
-    async def arun(
-        self,
-        prompt: str,
-        images: Image.Image | bytes | list[Image.Image | bytes] | None = None,
-        init_conversation: Optional[list[dict[str, str]]] = None,
-        image_resolution: str = "low",
-    ) -> CompletionResults:
-        messages = self.load_conversation()
-
-        if isinstance(init_conversation, list) and len(messages) == 0:
-            messages.extend(init_conversation)
-
-        messages.append(
-            Message(content=prompt, images=images, image_resolution=image_resolution).as_user
-        )
-
-        if self.content_filter is not None:
-            messages = self.apply_content_filter(messages)
-
-        messages_adjusted = self.conversation_length_adjuster.run(messages)
-
-        response = await self.chat_model.arun(messages_adjusted)
-
-        if self.content_filter is not None:
-            response.message = self.restore_content_filter([response.message])[0]
-
-        messages.append(response.message)
-
-        if self.conversation_memory is not None:
-            self.save_conversation(messages)
-
-        return response
+    def _get_client_message_type(self) -> type[BaseMessage]:
+        return OpenAIMessage
 
     async def abatch_run(
         self,
