+++ src/langrila/openai/model_config.py
